+++ sys/src/cmd/7a/a.h
@@ -120,7 +120,7 @@ EXTERN	int	sym;
 EXTERN	char	symb[NSYMB];
 EXTERN	int	thechar;
 EXTERN	char*	thestring;
-EXTERN	long	thunk;
+EXTERN	uintptr	thunk;
 EXTERN	Biobuf	obuf;
 
 void*	alloc(long);
